modular_code_replacement_for_crawler_py/requirements.txt
# Core Dependencies
crawl4ai>=0.7.8
google-genai>=1.0.0
pdfplumber>=0.11.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
requests>=2.31.0
tenacity>=8.0.0
aiohttp>=3.9.0

# Optional - for advanced PDF extraction
pymupdf>=1.24.0

# Optional - for OCR (requires Tesseract installed)
# pytesseract>=0.3.13
